data is coming and parsing fine
and all set up is ok for product details
