+++ lib/app/bindings/repository_bindings.dart
@@ -1,15 +1,22 @@
+import 'package:bs_commerce/app/data/product/repository/product_details/product_details_repository.dart';
+import 'package:bs_commerce/app/data/product/repository/product_details/product_details_repository_impl.dart';
 import 'package:get/get.dart';
 
-import '../data/product/repository/catalogue_repository.dart';
-import '../data/product/repository/catalogue_repository_impl.dart';
+import '../data/product/repository/home/home_repository.dart';
+import '../data/product/repository/home/home_repository_impl.dart';
 
 
 class RepositoryBindings implements Bindings {
   @override
   void dependencies() {
-    Get.lazyPut<CatalogueRepository>(
-      () => CatalogueRepositoryImpl(),
-      tag: (CatalogueRepository).toString(),
+    Get.lazyPut<HomeRepository>(
+      () => HomeRepositoryImpl(),
+      tag: (HomeRepository).toString(),
+    );
+    Get.lazyPut<ProductDetailsRepository>(
+          () => ProductDetailsRepositoryImpl(),
+      tag: (ProductDetailsRepository).toString(),
+      fenix: true
     );
   }
 }
