+++ lib/app/data/product/model/product_home/model_product_categories.dart
@@ -1,0 +1,19 @@
+class Categories {
+  String? id;
+  String? name;
+
+  Categories({id, name});
+
+  Categories.fromJson(Map<String, dynamic> json) {
+    id = json['id'];
+    name = json['name'];
+  }
+
+  Map<String, dynamic> toJson() {
+    final Map<String, dynamic> data = Map<String, dynamic>();
+    data['id'] = id;
+    data['name'] = name;
+
+    return data;
+  }
+}
