+++ lib/app/data/product/model/product_home/model_product_manufacturer.dart
@@ -1,0 +1,19 @@
+class Manufacturer {
+  String? id;
+  String? name;
+
+  Manufacturer({id, name});
+
+  Manufacturer.fromJson(Map<String, dynamic> json) {
+    id = json['id'];
+    name = json['name'];
+  }
+
+  Map<String, dynamic> toJson() {
+    final Map<String, dynamic> data = Map<String, dynamic>();
+    data['id'] = id;
+    data['name'] = name;
+
+    return data;
+  }
+}
