--- lib/app/data/product/remote/catalogue_remote_data_source.dart
@@ -1,6 +1,0 @@
-import '../model/data_model.dart';
-import '../model/product_catelog_response.dart';
-
-abstract class CatalogueRemoteDataSource {
-  Future<ProductCatalogResponse> getDetails();
-}
