+++ lib/app/data/product/remote/home/home_remote_data_source.dart
@@ -1,0 +1,5 @@
+import '../../model/product_home/product_home_response.dart';
+
+abstract class HomeRemoteDataSource {
+  Future<ProductHomeResponse> getHomePageProducts();
+}
