+++ lib/app/data/product/remote/product_details/product_details_remote_data_source.dart
@@ -1,0 +1,5 @@
+import 'package:bs_commerce/app/data/product/model/product_details/ProductDetailsResponse.dart';
+
+abstract class ProductDetailsRemoteDataSource {
+  Future<ProductDetailsResponse> getProductDetails(String productId);
+}
