--- lib/app/data/product/repository/catalogue_repository.dart
@@ -1,6 +1,0 @@
-import '../model/data_model.dart';
-import '../model/product_catelog_response.dart';
-
-abstract class CatalogueRepository {
-  Future<ProductCatalogResponse> getDetails();
-}
