+++ lib/app/data/product/repository/home/home_repository.dart
@@ -1,0 +1,6 @@
+
+import '../../model/product_home/product_home_response.dart';
+
+abstract class HomeRepository {
+  Future<ProductHomeResponse> getHomePageProducts();
+}
