+++ lib/app/data/product/repository/product_details/product_details_repository.dart
@@ -1,0 +1,5 @@
+import 'package:bs_commerce/app/data/product/model/product_details/ProductDetailsResponse.dart';
+
+abstract class ProductDetailsRepository {
+  Future<ProductDetailsResponse> getProductDetails(String productId);
+}
