+++ lib/app/modules/home/controllers/home_controller.dart
@@ -1,23 +1,23 @@
-import 'package:bs_commerce/app/data/product/model/product_catelog_response.dart';
 import 'package:get/get.dart';
 
-import '../../../data/product/repository/catalogue_repository.dart';
+import '../../../data/product/model/product_home/product_home_response.dart';
+import '../../../data/product/repository/home/home_repository.dart';
 import '/app/core/base/base_controller.dart';
 import '/app/modules/home/model/ui_data.dart';
 
 class HomeController extends BaseController {
-  final CatalogueRepository _repository =
-      Get.find(tag: (CatalogueRepository).toString());
+  final HomeRepository _repository =
+      Get.find(tag: (HomeRepository).toString());
 
   final Rx<UiData?> _data = Rx(null);
 
   UiData? get data => _data.value;
 
-  getDetails() {
-    callDataService(_repository.getDetails(), onSuccess: _handleDataResponse);
+  getHomePageProducts() {
+    callDataService(_repository.getHomePageProducts(), onSuccess: _handleDataResponse);
   }
 
-  _handleDataResponse(ProductCatalogResponse data) {
+  _handleDataResponse(ProductHomeResponse data) {
     _data(UiData(data));
   }
 }
