+++ lib/app/modules/home/model/ui_data.dart
@@ -1,7 +1,8 @@
-import 'package:bs_commerce/app/data/product/model/product_catelog_response.dart';
+
+import '../../../data/product/model/product_home/product_home_response.dart';
 
 class UiData {
-  final ProductCatalogResponse? value;
+  final ProductHomeResponse? value;
 
   UiData(this.value);
 }
