+++ lib/app/modules/product_details/model/product_details_ui_data.dart
@@ -1,0 +1,11 @@
+
+import 'package:bs_commerce/app/data/product/model/product_details/ProductDetailsResponse.dart';
+
+import '../../../data/product/model/product_home/product_home_response.dart';
+
+class ProductDetailsUiData {
+
+  final ProductDetailsResponse? value;
+
+  ProductDetailsUiData(this.value);
+}
